Generate CSV functionality for Posts
